Make array ranking more efficient for longer arrays with more distinct numbers.

For a longer array with a larger variety of numbers, `indexOf` and `lastIndexOf` take longer and are called more frequently. Eliminate the need to call `indexOf` and `lastIndexOf` by using the frequency of each number in the array.

**Sample times taken**
Array of length 100, with random numbers between 1 and 100
- `previous rank: 0.072021484375ms`
- `new rank: 0.054931640625ms`

Array of length 100,000, ascending distinct numbers from 0 through 99,999
- `previous rank: 14851.73291015625ms`
- `new rank: 33.7861328125ms`

Array of length 100,000, with random numbers between 1 and 100
- `previous rank: 14704.958984375ms`
- `new rank: 37.547119140625ms`

**Note that previous rank is slightly faster when all numbers in the array are the same:**
Array of length 1 million, all `1`s
- `previous rank: 53.232177734375ms`
- `new rank: 162.830078125ms`