@suzukia 2020-03-14
Hello!
Thank you for providing this useful library! I just have a suggestion that makes array ranking faster for most use cases.