@trevnorris 2020-03-26
@suzukia This is excellent. Thanks for the contribution. Can I ask what code you used to perform the benchmarks, and what you used to test it (e.g. chrome, node.js, etc.)?