@meenie 2020-04-01
> @suzukia This is excellent. Thanks for the contribution. Can I ask what code you used to perform the benchmarks, and what you used to test it (e.g. chrome, node.js, etc.)?

Hey @trevnorris, 

We tested this on Node.js as well and the numbers are pretty much identical.

We noticed that when pushing through large sets of numbers in the rank function that the time to finish was going up exponentially. I think at one point I was trying to run 500K numbers and it was taking 10 minutes to finish.  Using Ami's new ranking algorithm turned that into 500ms.